chore(gulp): refactor docs related tasks
