+++ package.json
@@ -27,7 +27,7 @@
     "docs:build": "npm-run-all docs:prepare docs:prod docs:dirs",
     "docs:prod": "npm run build -- docs --configuration production --aot --base-href '/nebular/'",
     "docs:dirs": "npm-run-all -s docs:assets:prod docs:copy-index docs:sitemap",
-    "docs:copy-index": "gulp create-docs-dirs",
+    "docs:copy-index": "gulp docs-copy-index",
     "docs:sitemap": "gulp create-sitemap",
     "docs:gh-pages": "ts-node -P ./scripts/docs/tsconfig.json ./scripts/docs/build-docs.ts",
     "build": "ng build",
@@ -139,6 +139,7 @@
     "fs-extra": "^8.1.0",
     "gulp": "4.0.2",
     "gulp-replace": "1.0.0",
+    "gulp-sass": "^5.0.0",
     "gulp-typedoc": "2.2.3",
     "husky": "^7.0.2",
     "jasmine-core": "~3.8.0",
