+++ tools/gulp/config.ts
@@ -1,11 +1,8 @@
 export const SOURCE_DIR = './src/framework';
-export const BUILD_DIR = './.ng_build';
-export const ES5_BUILD_DIR = './.es5_build';
 export const DIST_DIR = './dist';
 export const DOCS_OUTPUT = './docs/output.json';
 export const EXTENSIONS = ['ts', 'html', 'scss'];
 export const DOCS_DIST = './docs/dist';
 export const DEV_SCHEMATICS_PATH = 'tools/dev-schematics';
-export const JS_PACKAGES = ['auth', 'date-fns', 'eva-icons', 'moment', 'security', 'theme', 'firebase-auth'];
 export const DOCS_SITE_URL = 'https://akveo.github.io/nebular/';
 export const SCHEMATICS_SPEC_TSCONFIG = 'tsconfig.schematics.spec.json';
