+++ tools/gulp/docs/prepare-docs.ts
@@ -1,0 +1,6 @@
+import { task, series } from 'gulp';
+
+import { generateOuputJson } from './output-json';
+import { prepareExamples } from './example';
+
+task('docs', series(generateOuputJson, prepareExamples));
