+++ tools/gulp/docs/sass-export.ts
@@ -1,3 +1,4 @@
+import { join } from 'path';
 const _ = require('lodash');
 const fs = require('fs');
 const Colors = require('colors.js');
@@ -23,20 +24,53 @@ class PropLink {
   }
 }
 
-const exporter = {
+export function getExportFunction(outDir: string) {
+  const sef = new SassExportFunction(outDir);
+  return { 'export($file, $themes, $mapping)': sef.export.bind(sef) };
+}
+
+class SassExportFunction {
+  constructor(private outDir: string) {}
+
+  export(file, themes, mapping) {
+    const themesValue = this.get_value(themes);
+    const mappingValue = this.get_value(mapping);
+
+    const completeThemes = {};
+    Object.keys(themesValue).forEach((themeName) => {
+      const theme = themesValue[themeName];
+      completeThemes[themeName] = {
+        ...theme,
+        data: _.defaults(mappingValue, theme.data),
+      };
+    });
+
+    let output = {
+      themes: this.parseThemes(themesValue),
+      // TODO: we need to change internal function interface as it very hard to re-use them
+      completeThemes: this.parseThemes(completeThemes),
+    };
+
+    const outFilePath = join(this.outDir, file.getValue());
+    output = _.defaults(JSON.parse(fs.readFileSync(outFilePath)), output);
+    fs.writeFileSync(outFilePath, JSON.stringify(output, null, '  '));
+
+    return themes;
+  }
+
   get_value(a) {
     let value, i;
     switch (a.constructor.name) {
       case 'SassList':
         value = [];
         for (i = 0; i < a.getLength(); i++) {
-          value.push(exporter.get_value(a.getValue(i)));
+          value.push(this.get_value(a.getValue(i)));
         }
         break;
       case 'SassMap':
         value = {};
         for (i = 0; i < a.getLength(); i++) {
-          value[a.getKey(i).getValue()] = exporter.get_value(a.getValue(i));
+          value[a.getKey(i).getValue()] = this.get_value(a.getValue(i));
         }
         break;
       case 'SassColor':
@@ -59,7 +93,7 @@ const exporter = {
         value = a.getValue();
     }
     return value;
-  },
+  }
 
   parseThemes(THEMES) {
     let result = {};
@@ -73,12 +107,12 @@ const exporter = {
 
       Object.keys(theme).forEach((prop) => {
         result[themeName].data[prop] = result[themeName].data[prop] ? result[themeName].data[prop] : new Prop(prop);
-        result = exporter.getParent(prop, themeName, themeName, prop, result, THEMES);
+        result = this.getParent(prop, themeName, themeName, prop, result, THEMES);
       });
     });
 
     return result;
-  },
+  }
 
   getParent(prop, scopedThemeName, resultThemeName, resultProp, resultObj, THEMES) {
     const scopedTheme = THEMES[scopedThemeName].data;
@@ -86,23 +120,23 @@ const exporter = {
     const value = scopedTheme[prop];
     if (typeof value === 'string' && scopedTheme[value]) {
       if (resultObj[resultThemeName].data[resultProp].parents.length === 0) {
-        exporter.linkProps(resultObj, scopedThemeName, value, resultThemeName, prop);
+        this.linkProps(resultObj, scopedThemeName, value, resultThemeName, prop);
       } else {
         resultObj[resultThemeName].data[resultProp].parents.push(new PropLink(scopedThemeName, value));
       }
-      exporter.getParent(value, scopedThemeName, resultThemeName, resultProp, resultObj, THEMES);
+      this.getParent(value, scopedThemeName, resultThemeName, resultProp, resultObj, THEMES);
     } else {
       resultObj[resultThemeName].data[resultProp].value = value;
       if (scopedParent && THEMES[scopedParent].data[prop] === value) {
         if (resultObj[resultThemeName].data[resultProp].parents.length === 0) {
-          exporter.linkProps(resultObj, scopedParent, prop, resultThemeName, prop);
+          this.linkProps(resultObj, scopedParent, prop, resultThemeName, prop);
         } else {
           resultObj[resultThemeName].data[resultProp].parents.push(new PropLink(scopedParent, prop));
         }
       }
     }
     return resultObj;
-  },
+  }
 
   linkProps(resultObj, parentThemeName, parentPropName, childThemeName, childPropName) {
     if (!resultObj.hasOwnProperty(parentThemeName)) {
@@ -114,42 +148,5 @@ const exporter = {
     resultObj[childThemeName].data[childPropName].parents.push(new PropLink(parentThemeName, parentPropName));
     resultObj[parentThemeName].data[parentPropName].childs.push(new PropLink(childThemeName, childPropName));
     return resultObj;
-  },
-
-  function(path) {
-    return function (file, themes, mapping) {
-      const themesValue = exporter.get_value(themes);
-      const mappingValue = exporter.get_value(mapping);
-
-      const completeThemes = {};
-      Object.keys(themesValue).forEach((themeName) => {
-        const theme = themesValue[themeName];
-        completeThemes[themeName] = {
-          ...theme,
-          data: _.defaults(mappingValue, theme.data),
-        };
-      });
-
-      let output = {
-        themes: exporter.parseThemes(themesValue),
-        // TODO: we need to change internal function interface as it very hard to re-use them
-        completeThemes: exporter.parseThemes(completeThemes),
-      };
-
-      output = _.defaults(JSON.parse(fs.readFileSync(path + '/' + file.getValue())), output);
-      fs.writeFileSync(path + '/' + file.getValue(), JSON.stringify(output, null, '  '));
-      return themes;
-    };
-  },
-
-  interface(name) {
-    name = name || 'export';
-    return name + '($file, $themes, $mapping)';
-  },
-};
-
-export function exportThemes(path, name) {
-  const out = {};
-  out[exporter.interface(name)] = exporter.function(path);
-  return out;
+  }
 }
