+++ tools/gulp/gulpfile.ts
@@ -6,3 +6,6 @@
 
 import './schematics/build-dev-schematics';
 import './schematics/test-packages-schematics';
+import './docs/prepare-docs';
+import './docs/copy-index';
+import './docs/sitemap';
